requirements.txt
# Requisitos para el proyecto de Clima y Noticias de Chile
# Generado automáticamente

# Bibliotecas principales
requests==2.31.0
beautifulsoup4==4.13.4
python-dateutil==2.9.0.post0

# Biblioteca para noticias chilenas
chilescrapper==0.1.2

# Bibliotecas para Gmail API (si se implementa el envío real)
google-api-python-client==2.107.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Utilidades
PyYAML==6.0.2
urllib3==2.4.0
